+++ tests/test_create_observation/test_create_property_popularity_observation.py
@@ -2,31 +2,14 @@
 
 from urllib.error import HTTPError
 import pytest
-from fetch_data import (
-    create_property_popularity_observation,
-    update_out_of_date_property_observations,
-)
+from fetch_data import create_property_popularity_observation
 
 
 @pytest.mark.asyncio
 @pytest.mark.dependency(
-    name="property-popularity-success-ood", depends=["add-wikibase"], scope="session"
-)
-@pytest.mark.property
-@pytest.mark.sparql
-async def test_update_out_of_date_property_observations_success(mocker):
-    """Test One-Pull Per Month, Data Returned Scenario"""
-
-    mocker.patch(
-        "fetch_data.sparql_data.create_property_popularity_data_observation.get_sparql_results",
-        side_effect=[{"results": {"bindings": []}}],
-    )
-    await update_out_of_date_property_observations()
-
-
-@pytest.mark.asyncio
-@pytest.mark.dependency(
-    name="property-popularity-success", depends=["add-wikibase"], scope="session"
+    name="property-popularity-success",
+    depends=["property-popularity-success-ood"],
+    scope="session",
 )
 @pytest.mark.property
 @pytest.mark.sparql
