+++ tests/test_create_observation/test_create_quantity_observation.py
@@ -3,31 +3,12 @@
 import time
 from urllib.error import HTTPError
 import pytest
-from fetch_data import (
-    create_quantity_observation,
-    update_out_of_date_quantity_observations,
-)
+from fetch_data import create_quantity_observation
 
 
 @pytest.mark.asyncio
 @pytest.mark.dependency(
-    name="quantity-success-ood", depends=["add-wikibase"], scope="session"
-)
-@pytest.mark.quantity
-@pytest.mark.sparql
-async def test_update_out_of_date_quantity_observations_success(mocker):
-    """Test"""
-
-    mocker.patch(
-        "fetch_data.sparql_data.create_quantity_data_observation.get_sparql_results",
-        side_effect=[],
-    )
-    await update_out_of_date_quantity_observations()
-
-
-@pytest.mark.asyncio
-@pytest.mark.dependency(
-    name="quantity-success", depends=["add-wikibase"], scope="session"
+    name="quantity-success", depends=["quantity-success-ood"], scope="session"
 )
 @pytest.mark.quantity
 @pytest.mark.sparql
@@ -48,7 +29,9 @@ async def test_create_quantity_observation_success(mocker):
 
 
 @pytest.mark.asyncio
-@pytest.mark.dependency(name="quantity-failure", depends=["quantity-success"])
+@pytest.mark.dependency(
+    name="quantity-failure", depends=["quantity-success"], scope="session"
+)
 @pytest.mark.quantity
 @pytest.mark.sparql
 async def test_create_quantity_observation_failure(mocker):
