+++ tests/test_create_observation/test_create_software_version_observation/test_create_software_version_observation.py
@@ -4,10 +4,7 @@ import os
 import time
 from urllib.error import HTTPError
 import pytest
-from fetch_data import (
-    create_software_version_observation,
-    update_out_of_date_software_observations,
-)
+from fetch_data import create_software_version_observation
 from tests.test_create_observation.test_create_software_version_observation.test_constants import (
     DATA_DIRECTORY,
 )
@@ -17,23 +14,9 @@ from tests.utils import MockResponse
 
 @pytest.mark.asyncio
 @pytest.mark.dependency(
-    name="software-version-success-ood", depends=["add-wikibase"], scope="session"
-)
-@pytest.mark.soup
-@pytest.mark.version
-async def test_update_out_of_date_software_observations_fail(mocker):
-    """Test Data Returned Scenario"""
-
-    mocker.patch(
-        "fetch_data.soup_data.software.create_software_version_data_observation.requests.get",
-        side_effect=[MockResponse("", 404)],
-    )
-    await update_out_of_date_software_observations()
-
-
-@pytest.mark.asyncio
-@pytest.mark.dependency(
-    name="software-version-success", depends=["add-wikibase"], scope="session"
+    name="software-version-success",
+    depends=["software-version-fail-ood"],
+    scope="session",
 )
 @pytest.mark.soup
 @pytest.mark.version
@@ -58,7 +41,9 @@ async def test_create_software_version_observation_success(mocker):
 
 @pytest.mark.asyncio
 @pytest.mark.dependency(
-    name="software-version-failure", depends=["software-version-success"]
+    name="software-version-failure",
+    depends=["software-version-success"],
+    scope="session",
 )
 @pytest.mark.soup
 @pytest.mark.version
