+++ tests/test_create_observation/test_create_statistics_observation/test_create_statistics_observation.py
@@ -4,10 +4,7 @@ import os
 import time
 from urllib.error import HTTPError
 import pytest
-from fetch_data import (
-    create_special_statistics_observation,
-    update_out_of_date_stats_observations,
-)
+from fetch_data import create_special_statistics_observation
 from tests.utils import MockResponse
 
 
@@ -16,23 +13,7 @@ DATA_DIRECTORY = "tests/test_create_observation/test_create_statistics_observati
 
 @pytest.mark.asyncio
 @pytest.mark.dependency(
-    name="statistics-success-ood", depends=["add-wikibase"], scope="session"
-)
-@pytest.mark.soup
-@pytest.mark.statistics
-async def test_update_out_of_date_stats_observations_success(mocker):
-    """Test Data Returned Scenario"""
-
-    mocker.patch(
-        "fetch_data.soup_data.create_statistics_data_observation.requests.get",
-        side_effect=[MockResponse("", 404)],
-    )
-    await update_out_of_date_stats_observations()
-
-
-@pytest.mark.asyncio
-@pytest.mark.dependency(
-    name="statistics-success", depends=["add-wikibase"], scope="session"
+    name="statistics-success", depends=["statistics-fail-ood"], scope="session"
 )
 @pytest.mark.soup
 @pytest.mark.statistics
@@ -52,7 +33,9 @@ async def test_create_statistics_observation_success(mocker):
 
 
 @pytest.mark.asyncio
-@pytest.mark.dependency(name="statistics-failure", depends=["statistics-success"])
+@pytest.mark.dependency(
+    name="statistics-failure", depends=["statistics-success"], scope="session"
+)
 @pytest.mark.soup
 @pytest.mark.statistics
 async def test_create_statistics_observation_failure(mocker):
