+++ tests/test_create_observation/test_ood/test_ood_user_observation.py
@@ -1,0 +1,19 @@
+"""Test create_user_observation"""
+
+import pytest
+from fetch_data import update_out_of_date_user_observations
+
+
+@pytest.mark.asyncio
+@pytest.mark.dependency(
+    name="user-empty-ood", depends=["add-wikibase"], scope="session"
+)
+@pytest.mark.user
+async def test_update_out_of_date_user_observations_empty(mocker):
+    """Test No-Data Scenario"""
+
+    mocker.patch(
+        "fetch_data.api_data.user_data.fetch_all_user_data.fetch_api_data",
+        side_effect=[{"query": {"allusers": []}}],
+    )
+    await update_out_of_date_user_observations()
